kereső mező lessgoo és törlés

+++ pages/api/posts/delete.ts
@@ -1,5 +1,5 @@
+import getUser from "@lib/server/getUser";
 import prisma from "@lib/server/prisma";
-import { parseBody } from "next/dist/next-server/server/api-utils";
 import wrapper from "../../../lib/server/endpoint";
 
 type DeletePostRequiredParameters = {
@@ -7,7 +7,22 @@ type DeletePostRequiredParameters = {
 };
 
 export default wrapper(async (req) => {
+    const user = await getUser(req);
     const { id }: DeletePostRequiredParameters = JSON.parse(req.body);
+
+    const post = await prisma.post.findUnique({
+        where: {
+            id,
+        },
+    });
+
+    if(post.userId !== user.id) throw new Error("Nincs jogosultságod ehhez!");
+
+    await prisma.postCategoryConnector.deleteMany({
+        where: {
+            postId: id,
+        },
+    });
 
     await prisma.post.delete({
         where: {
